requirements.txt
flask
python-docx
Jinja2
pytest
pyyaml
google-cloud-aiplatform>=1.59.0
google-cloud-logging
firebase-admin
langchain-google-vertexai>=1.0.7